requirements.txt
google-ai-generativelanguage==0.4.0
google-generativeai==0.4.0
ipython==8.22.2
numpy==1.26.4
pandas==2.2.0
python-docx==1.1.0
streamlit==1.31.1
